refactor: de-part region map component support for wave 9 slice E (#4117)
Extract types, session state, lifecycle, and interaction from the
region_map_component part library into region_map_component_support.dart;
shrink the no-part grandfather by three entries while staying within the
main-line budget.

Refs #4117
