+++ app/lib/features/game/flame/region_map/region_map_component_render_markers_settlements_warp.dart
@@ -109,7 +109,7 @@ extension _CtRegionMapRenderMarkersSettlementsWarp on CtRegionMapComponent {
     }
 
     if (!regionMapDrawBoundaryBetweenAdjacentCells(
-      gateByUnrevealedTiles: _gateMapBoundariesByVisibility,
+      gateByUnrevealedTiles: gateMapBoundariesByVisibility,
       visibilityA: _visibilityForTerrain(cell),
       visibilityB: _visibilityForTerrain(neighbor),
     )) {
