+++ app/lib/features/game/flame/region_map/region_map_component_render_markers_units_civilian.dart
@@ -25,7 +25,7 @@ extension _CtRegionMapRenderMarkersUnitsCivilian on CtRegionMapComponent {
           ? _kHoveredProvinceGlowOpacityMid +
                 _kHoveredProvinceGlowOpacityAmplitude *
                     math.sin(
-                      _hoverAnimationT * _kHoveredProvinceGlowAngularFrequency,
+                      session.hoverAnimationT * _kHoveredProvinceGlowAngularFrequency,
                     )
           : 1.0;
