+++ app/lib/features/game/flame/region_map/region_map_component_render_political_labels_province_paint.dart
@@ -5,7 +5,7 @@ extension _CtRegionMapRenderPoliticalLabelsProvincePaint
     on CtRegionMapComponent {
   void _paintProvinceNames(Canvas canvas) {
     _ensureProvinceLabelCache();
-    final labels = _provinceLabelsCached;
+    final labels = session.provinceLabelsCached;
     if (labels == null || labels.isEmpty) {
       return;
     }
